requirements.txt


Django>=5.0,<6
djangorestframework>=3.15,<4
djangorestframework-simplejwt>=5.3,<6
django-cors-headers>=4.3,<5
psycopg2-binary>=2.9,<3
python-dotenv>=1.0,<2

# Etapa 4 — PDFs, embeddings (Ollama), ChromaDB
pypdf>=4.0,<6
httpx>=0.27,<1
chromadb>=0.5,<1

# Processamento assíncrono de uploads (Celery + Redis em produção; eager em testes)
celery[redis]>=5.3,<6
redis>=5.0,<6
